Allow specifying list of source file search paths in environment variable.
This is useful for binaries (using backward-cpp) that are built with
relative source file paths.  In such a situation, backward-cpp will
generally not be able to locate the source files.  This change allows
the user to specify a delimited list of search paths in the following
environment variable:

  BACKWARD_CXX_SOURCE_PREFIXES

The paths are separated by semicolons on Windows and colons otherwise.
When the environment variable is set, those search paths will be tried
first.  If they do not yield any valid files then backward-cpp will
fall back to current behavior.